+++ src/Bus/CommandHandlerInterface.php
@@ -20,6 +20,7 @@ declare(strict_types=1);
 namespace CloudCreativity\Modules\Bus;
 
 use CloudCreativity\Modules\Toolkit\Messages\CommandInterface;
+use CloudCreativity\Modules\Toolkit\Messages\DispatchThroughMiddleware;
 use CloudCreativity\Modules\Toolkit\Result\ResultInterface;
 
 interface CommandHandlerInterface extends DispatchThroughMiddleware